cabin-display/pyproject.toml
[project]
name = "cabine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pyqt6>=6.9.1",
]
